# api/openapi.yaml
openapi: 3.0.0
info:
  title: Matrix Computation API
  description: API to compute a matrix of double values from input parameters.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local server
  - url: http://azakan.net:5000
    description: Remote server
paths:
...
  /api/zabr_calibration:
    get:
      summary: Get ZABR calibration results
      operationId: getZabrCalibration
      x-swagger-router-controller: Default
      parameters:
        - name: N
          in: query
          required: false
          schema:
            type: integer
            default: 401
            description: Number of grid points for volatility surface
        - name: forward
          in: query
          required: false
          schema:
            type: number
            default: 0.02
            description: Forward price
        - name: expiry
          in: query
          required: false
          schema:
            type: number
            default: 30
            description: Time to expiry
        - name: alpha
          in: query
          required: false
          schema:
            type: number
            default: 0.00955
            description: Alpha parameter
        - name: beta
          in: query
          required: false
          schema:
            type: number
            default: 0.956
            description: Beta parameter
        - name: vol_of_vol
          in: query
          required: false
          schema:
            type: number
            default: 0.373
            description: Volatility of volatility
        - name: rho
          in: query
          required: false
          schema:
            type: number
            default: -0.749
            description: Correlation parameter
        - name: shift
          in: query
          required: false
          schema:
            type: number
            default: 0.0
            description: Shift parameter
        - name: gamma
          in: query
          required: false
          schema:
            type: number
            default: 1.0
            description: Gamma parameter
        - name: dt
          in: query
          required: false
          schema:
            type: number
            default: 5.0
            description: dt parameter
        - name: nd
          in: query
          required: false
          schema:
            type: number
            default: 3.5
            description: nd parameter
        - name: calibration_type
          in: query
          required: false
          schema:
            type: string
            enum: [classical, pde, mixture]
            default: classical
            description: Type of calibration to perform
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success, error]
                  data:
                    type: object
                    properties:
                      strikes:
                        type: array
                        items:
                          type: number
                      model_vols:
                        type: array
                        items:
                          type: number
                      market_strikes:
                        type: array
                        items:
                          type: number
                      market_vols:
                        type: array
                        items:
                          type: number
                  error:
                    type: string
                    nullable: true
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - status
        - error
      properties:
        status:
          type: string
          enum: [error]
          description: Error status indicator
        error:
          type: string
          description: Detailed error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp of when the error occurred
...